Add the utility class JImagePanel to integrate better with Swing
When you want to show an ImagePlus in a Swing-based window, ImageCanvas
is not the best choice, as AWT and Swing do not mix well.

FocalPoint has a public domain version of a Swing component that can
show an ImagePlus:

	http://article.gmane.org/gmane.comp.java.imagej/14023

Include that, so that Fiji's plugins can benefit from it.

Here is an example how to use it with Jython:

	from fiji.utilities.gui import JImagePanel
	from javax.swing import JDialog

	IJ.run("Clown (14K)");

	panel = JImagePanel(WindowManager.getCurrentImage())
	dialog = JDialog(None, "JImagePanel test")
	dialog.getContentPane().add(panel);
	dialog.pack()
	dialog.show()

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>