Use ImageJ2's sample uploader in Help>Upload Sample Image
With the recent meltdown of the server that used to serve fiji.sc,
we needed to change a couple of things, including the sample image
uploading: the new fiji.sc server is behind a firewall and cannot be
accessed via SSH.

Happily, we did a lot of work there already and implemented a
WebDAV-based uploader in ImageJ2. Now all we need to do is to use
that uploader in Fiji, which is what this commit does.

Requested by Jean-Yves Tinevez.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>