Remove obsolete Fiji_Uploader plugin
This plugin moved to ImageJ2 quite some time ago:
https://github.com/imagej/imagej-plugins-commands/blob/imagej-plugins-commands-0.4.1/src/main/java/net/imagej/plugins/commands/upload/SampleImageUploader.java

In most circumstances, we would try harder to maintain backwards
compatibility. But in this case, Fiji_Uploader was dragging in a
dependency on fiji-compat, which drags in a huge dependency tree.
With this change, the dependency list goes from 91 down to 4.