Keep extreme distance values to adjust contrast
When running the SEDT plugin on a stack, the display range is only
adjusted for the first picture of the stack. In order to obtain
unsaturated pictures for the whole stack, I needed to pass the whole
stack through the "Enhance Contrast" process.

To limit stack IO operations, the extreme values of distance are kept
during the SEDT process in order to have the correct display range in
the resulting image.

This is done by modifying the X::set() method, since it is the last
set() method to be called in the plugin.

One drawback of using this method instead of the "Enhance Contrast"
method is that here it is implied that the "Saturated pixels" value is
0.0% (since the absolute minimum and maximum are used). But a Euclidean
Distance Transform should not give "extreme" values for which having
absolute minimum and maximum is a problem.

Closes #4.

Signed-off-by: Curtis Rueden <ctrueden@wisc.edu>